requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
streamlit
pandas
python-dotenv
pypdf
